+++ src/pages/mypage/MyPage.tsx
@@ -1,27 +1,12 @@
-import { useState } from 'react';
 import ProfileSection from './components/profile/ProfileSection';
 import HistorySection from './components/history/HistorySection';
 import SettingSection from './components/setting/SettingSection';
 import * as styles from './MyPage.css';
-import { useToast } from '@/shared/components/toast/useToast';
 import TitleNavBar from '@/shared/components/navBar/TitleNavBar';
 
 const MyPage = () => {
-  const userName = '';
+  const userName = '하우미';
   const credit = 0;
-  const { notify } = useToast();
-  const [isChargeDisabled, setIsChargeDisabled] = useState(false);
-
-  const handleChargeClick = () => {
-    setIsChargeDisabled(true);
-    notify({
-      text: '결제는 아직 준비 중인 기능이에요',
-      type: 'warning',
-      options: {
-        style: { marginBottom: '2rem' },
-      },
-    });
-  };
 
   return (
     <div className={styles.contentWrapper}>
@@ -29,8 +14,7 @@ const MyPage = () => {
       <ProfileSection
         userName={userName}
         credit={credit}
-        onChargeClick={handleChargeClick}
-        isChargeDisabled={isChargeDisabled}
+        isChargeDisabled={false}
       />
       <HistorySection />
       <SettingSection />
